requirements.txt
#opencv-python~=4.11.0.86


einops~=0.8.1
timm~=1.0.15
wandb~=0.19.9
python-dotenv~=1.1.0
numpy~=2.2.3
tqdm~=4.67.1
#torchvision~=0.21.0
pillow~=11.1.0
kagglehub~=0.3.11
PyWavelets~=1.8.0
pytorch_msssim
focal_frequency_loss
matplotlib